resolve circulur dependencies by recombining firetask, workflow, and firework into firework module
